+++ src/ScottPlot5/ScottPlot5/Control/StandardActions.cs
@@ -121,7 +121,7 @@ public static class StandardActions
 
     public static void ToggleBenchmark(IPlotControl control)
     {
-        control.Plot.ShowBenchmark = !control.Plot.ShowBenchmark;
+        control.Plot.Benchmark.IsVisible = !control.Plot.Benchmark.IsVisible;
         control.Refresh();
     }